+++ DAL/Models/DishDto.cs
@@ -7,10 +7,10 @@ namespace WebApplication2.DAL.Models
             public Guid Id { get; set; }
             public string Name { get; set; }
             public string Description { get; set; }
-            public int Price { get; set; }
+            public double Price { get; set; }
             public string Image { get; set; }
             public bool Vegetarian { get; set; }
-            public int Rating { get; set; }
+            public double Rating { get; set; }
             public DishCategory Category { get; set; }
       
     }
